int to floats for gpu chunk gen

+++ src/chunkGen.h
@@ -1,1 +1,1 @@
-int* chunkDataKernel(int chunkZ, int chunkX, int* heightMap, double* gradients);
+float* chunkDataKernel(int chunkZ, int chunkX, int* heightMap, double* gradients);
